feat: added logging and shared module
