+++ src/RecLambdaFunction/index.py
@@ -1,6 +1,7 @@
 import os
 import boto3
 from handler import handle_request
+import logging
 
 dynamodb = boto3.resource('dynamodb')
 table = dynamodb.Table(os.environ['DYNAMO_TABLE_NAME'])
@@ -9,4 +10,7 @@ debug = os.getenv('ENABLE_VERBOSE_AND_SENSITIVE_LOGGING', 'false') == 'true'
 def handler(event, context):
     if debug:
         print(event)
-    return handle_request(event, table)
+    try:
+        return handle_request(event, table)
+    except Exception as e:
+        logging.error(e)
